refactor(scan): extract models into separate files
