+++ packages/scan/src/index.ts
@@ -1,5 +1,6 @@
+export * from './models';
 export * from './Scan';
 export * from './Scans';
-export * from './ScanFactory ';
-export * from './HarEntryBuilder';
-export * from './commands';
+export * from './ScanFactory';
+export * from './ScanSettings';
+export * from './Target';
