+++ packages/scan/src/models/AttackParamLocation.ts
@@ -1,0 +1,9 @@
+export enum AttackParamLocation {
+  ARTIFICAL_FRAGMENT = 'artifical-fragment',
+  ARTIFICAL_QUERY = 'artifical-query',
+  BODY = 'body',
+  FRAGMENT = 'fragment',
+  HEADER = 'header',
+  PATH = 'path',
+  QUERY = 'query'
+}
