+++ packages/scan/src/models/Discovery.ts
@@ -1,0 +1,5 @@
+export enum Discovery {
+  CRAWLER = 'crawler',
+  ARCHIVE = 'archive',
+  OAS = 'oas'
+}
