+++ packages/scan/src/models/HttpMethod.ts
@@ -1,0 +1,17 @@
+export enum HttpMethod {
+  GET = 'GET',
+  PUT = 'PUT',
+  POST = 'POST',
+  PATCH = 'PATCH',
+  DELETE = 'DELETE',
+  COPY = 'COPY',
+  HEAD = 'HEAD',
+  OPTIONS = 'OPTIONS',
+  LINK = 'LINK',
+  UNLINK = 'UNLINK',
+  PURGE = 'PURGE',
+  LOCK = 'LOCK',
+  UNLOCK = 'UNLOCK',
+  PROPFIND = 'PROPFIND',
+  VIEW = 'VIEW'
+}
