+++ packages/scan/src/models/Issue.ts
@@ -1,0 +1,24 @@
+export interface Issue {
+  id: string;
+  scanId: string;
+  order: number;
+  solved: boolean;
+  details: string;
+  name: string;
+  severity: string;
+  protocol: string;
+  remedy: string;
+  exposure: string;
+  cvss: string;
+  cwe: string;
+  time: Date;
+  screenshots: unknown[];
+  originalRequest: Record<string, unknown>;
+  request: Record<string, unknown>;
+  frames: unknown[];
+  originalFrames: [];
+  response: Record<string, unknown>;
+  assignees: unknown[];
+  resources: string[];
+  comments: string[];
+}
