+++ packages/scan/src/models/IssueGroup.ts
@@ -1,0 +1,6 @@
+import { Severity } from './Severity';
+
+export interface IssueGroup {
+  number: number;
+  type: Severity;
+}
