+++ packages/scan/src/models/Module.ts
@@ -1,0 +1,4 @@
+export enum Module {
+  DAST = 'dast',
+  FUZZER = 'fuzzer'
+}
