+++ packages/scan/src/models/ScanStatus.ts
@@ -1,0 +1,11 @@
+export enum ScanStatus {
+  FAILED = 'failed',
+  DISRUPTED = 'disrupted',
+  RUNNING = 'running',
+  STOPPED = 'stopped',
+  QUEUED = 'queued',
+  SCHEDULED = 'scheduled',
+  PENDING = 'pending',
+  DONE = 'done',
+  PAUSED = 'paused'
+}
