+++ packages/scan/src/models/Severity.ts
@@ -1,0 +1,5 @@
+export enum Severity {
+  MEDIUM = 'Medium',
+  HIGH = 'High',
+  LOW = 'Low'
+}
